+++ core-ng-mongo-test/src/test/java/core/framework/mongo/module/TestMongoConfigTest.java
@@ -1,7 +1,5 @@
 package core.framework.mongo.module;
 
-import core.framework.mongo.impl.MockMongo;
-import core.framework.mongo.impl.MongoImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -19,8 +17,7 @@ class TestMongoConfigTest {
     }
 
     @Test
-    void createMongo() {
-        MongoImpl mongo = config.createMongo();
-        assertThat(mongo).isInstanceOf(MockMongo.class);
+    void connectionString() {
+        assertThat(config.connectionString(null).getDatabase()).isNotNull();
     }
 }